+++ src/arch/x86/mod.rs
@@ -34,6 +34,14 @@ mod x86_mem;
 ))]
 pub(crate) use x86_mem::_memmem_impl;
 
+mod x86_nechr;
+
+#[cfg(all(
+    any(target_arch = "x86_64", target_arch = "x86"),
+    any(target_feature = "sse2", target_feature = "avx")
+))]
+pub(crate) use x86_nechr::_memnechr_impl;
+
 mod x86_rchr;
 
 #[cfg(all(
@@ -41,6 +49,14 @@ mod x86_rchr;
     any(target_feature = "sse2", target_feature = "avx")
 ))]
 pub(crate) use x86_rchr::_memrchr_impl;
+
+mod x86_rnechr;
+
+#[cfg(all(
+    any(target_arch = "x86_64", target_arch = "x86"),
+    any(target_feature = "sse2", target_feature = "avx")
+))]
+pub(crate) use x86_rnechr::_memrnechr_impl;
 
 mod x86_rmem;
 
@@ -73,8 +89,14 @@ pub use x86_cpy::_memcpy_sse2;
 //pub use x86_mem::_memmem_avx;
 pub use x86_mem::_memmem_sse2;
 
+pub use x86_nechr::_memnechr_avx;
+pub use x86_nechr::_memnechr_sse2;
+
 pub use x86_rchr::_memrchr_avx;
 pub use x86_rchr::_memrchr_sse2;
+
+pub use x86_rnechr::_memrnechr_avx;
+pub use x86_rnechr::_memrnechr_sse2;
 
 //pub use x86_rmem::_memrmem_avx;
 pub use x86_rmem::_memrmem_sse2;