+++ src/arch/x86/x86_mem.rs
@@ -18,12 +18,14 @@ fn _memmem_basic(haystack: &[u8], needle: &[u8]) -> Option<usize> {
 
 #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
 #[target_feature(enable = "sse2")]
+#[allow(clippy::missing_safety_doc)]
 pub unsafe fn _memmem_sse2(haystack: &[u8], needle: &[u8]) -> Option<usize> {
     _memmem_sse2_impl(haystack, needle)
 }
 
 #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
 #[target_feature(enable = "avx")]
+#[allow(clippy::missing_safety_doc)]
 pub unsafe fn _memmem_avx(haystack: &[u8], needle: &[u8]) -> Option<usize> {
     _memmem_avx_impl(haystack, needle)
 }